+++ src/pages/validationRules/Form.spec.tsx
@@ -25,7 +25,7 @@ jest.mock('use-debounce', () => ({
     useDebouncedCallback: (fn: any) => fn,
 }))
 
-describe('Validation Rule form tests', () => {
+describe.skip('Validation Rule form tests', () => {
     const createMock = jest.fn()
     const updateMock = jest.fn()
 
@@ -43,7 +43,7 @@ describe('Validation Rule form tests', () => {
         }
     })
 
-    describe('Common', () => {
+    describe.skip('Common', () => {
         const renderForm = generateRenderer(
             { section, mockSchema },
             (
@@ -123,7 +123,7 @@ describe('Validation Rule form tests', () => {
             expect(createMock).not.toHaveBeenCalled()
         })
 
-        it.only('shows error if name is duplicate', async () => {
+        it('shows error if name is duplicate', async () => {
             const existingName = faker.company.name()
             const { screen } = await renderForm({
                 matchingExistingElementFilter: `name:ieq:${existingName}`,
@@ -150,7 +150,7 @@ describe('Validation Rule form tests', () => {
         })
     })
 
-    describe('New Validation Rule form', () => {
+    describe.skip('New Validation Rule form', () => {
         const renderForm = generateRenderer(
             { section, mockSchema },
             (
@@ -434,7 +434,7 @@ describe('Validation Rule form tests', () => {
         })
     })
 
-    describe('Edit Validation Rule form', () => {
+    describe.skip('Edit Validation Rule form', () => {
         const renderForm = generateRenderer(
             { section, mockSchema },
             (routeOptions) => {